Constants and redis client init moved to /shared/files. Added a timeout to prod-task script
