+++ api/fetch-products.js
@@ -1,31 +1,11 @@
 const query = require('../db/index.js');
-const dotenv = require('dotenv');
 const fetch = require('node-fetch');
 const fetchManufacturerAvailability = require('./fetch-manufacturer-availability.js');
 const updateAvailability = require('../db/update-availability.js');
 const insertProduct = require('../db/insert-product.js');
 const deleteProduct = require('../db/delete-product.js');
-const { promisify } = require("util");
-
-if (process.env.NODE_ENV !== 'production') dotenv.config();
-
-// Init Redis
-const redis_url = process.env.REDIS_URL || null;
-const redis = require("redis");
-const client = redis.createClient(redis_url);
-
-client.on("error", (error) => console.error(error));
-const getAsync = promisify(client.get).bind(client);
-const cacheTimer = process.env.CACHE_TIMER || 300;
-
-// Get value from Redis that depends upon an async call
-const getResult = async (manufacturer) => {
-  try {
-    return await getAsync(manufacturer);
-  } catch (err) {
-    console.log(err);
-  };
-};
+const { getResult, client } = require("../shared/init-redis-client.js");
+const { PRODUCT_URL, CACHE_TIMER } = require("../shared/constants.js");
 
 const processColors = (colorArray) => {
   let colors = "";
@@ -43,7 +23,7 @@ const fetchProducts = async (product) => {
   let productIDs = [];
   let manufacturers = [];
 
-  const url = `${process.env.PRODUCT_URL}${product}`; // Build URL
+  const url = `${PRODUCT_URL}${product}`; // Build URL
   let data;
   try {
     // Try to get the data
@@ -91,7 +71,7 @@ const fetchProducts = async (product) => {
           manufacturersFetched['manufacturer-list'].push(manufacturer);
           client.set('manufacturer-list', JSON.stringify({
             'manufacturer-list': manufacturersFetched['manufacturer-list']
-          }), 'EX', cacheTimer);
+          }), 'EX', CACHE_TIMER);
         };
       };
     } else if (await !Array.isArray(data.response)) {
