+++ db/get-product-items.js
@@ -1,29 +1,7 @@
 const query = require("./index.js");
 const format = require("pg-format");
-const { promisify } = require("util");
-const dotenv = require("dotenv");
-
-if (process.env.NODE_ENV !== "production") {
-  dotenv.config();
-}
-
-// Redis
-const redis_url = process.env.REDIS_URL || null;
-const redis = require("redis");
-const client = redis.createClient(redis_url);
-
-client.on("error", (error) => console.error(error));
-const getAsync = promisify(client.get).bind(client);
-
-// Response
-
-const getResult = async (req) => {
-  try {
-    return await getAsync(req.header("X-PRODUCT"));
-  } catch (err) {
-    console.log(err);
-  }
-};
+const { getResult, client } = require("../shared/init-redis-client.js");
+const { CACHE_TIMER } = require("../shared/constants.js");
 
 const getProductItems = async (req, res) => {
   // Get product's stored hash if available
@@ -32,7 +10,6 @@ const getProductItems = async (req, res) => {
 
   if (!result) {
     // If no stored hash, get it from the DB
-    const cacheTimer = process.env.CACHE_TIMER || 300;
     let queryString = format("SELECT * FROM %I", req.header("X-PRODUCT"));
     result = await query(queryString);
 
@@ -42,7 +19,7 @@ const getProductItems = async (req, res) => {
       req.header("X-PRODUCT"),
       JSON.stringify(resValue),
       "EX",
-      cacheTimer
+      CACHE_TIMER
     );
   } else {
     resValue = result;
