+++ jobs/prod-task.js
@@ -1,6 +1,15 @@
 const fetchProducts = require("../api/fetch-products.js");
+const subscriberInit = require("../shared/subscriber-init.js");
+
+subscriberInit();
 
 const products = ["beanies", "facemasks", "gloves"];
 products.forEach((product, index) =>
   setTimeout(fetchProducts, 5000 * index, product)
 );
+
+// Kill script if it hasn't killed itself
+setTimeout((function() {
+  console.log('timeout');
+  return process.exit(22);
+}), 120000);
